+++ src/app/academy/(workspace)/[id]/dashboard/components/(principal)/ChallengeDetail/index.tsx
